simplify

+++ run_example_server.sh
@@ -1,8 +1,7 @@
 #!/bin/bash
 
-ETH=$(dmesg | grep -Eo 'eth[[:digit:]]+' | tail -n1)
-IP=$(ifconfig ${ETH} | head -n2 | tail -n1 | cut -d: -f2 | cut -d" " -f1)
-#IP=127.0.0.1
+ADDR=127.0.0.1:8000
+#ADDR=0.0.0.0:8000
 
 CMD=runserver
 #echo -n "Use 'runserver_plus' [y/n]: "
@@ -15,10 +14,15 @@ CMD=runserver
 
 cd example_project
 
+TEST_DB=example_project_db.sqlite3 # must be the same as in settings.py !
+
 echo
 (
+    if [ -f ${TEST_DB} ]; then
+        # remove old
+        rm example_project_db.sqlite3
+    fi
     set -x
-    rm example_project_db.sqlite3
     ./manage.py migrate
     echo "from django.contrib.auth.models import User; User.objects.create_superuser('test', 'test@test.com', '12345678')" | ./manage.py shell
 )
@@ -27,7 +31,7 @@ while true
 do
 (
     set -x
-    ./manage.py ${CMD} ${IP}:8000 --insecure
+    ./manage.py ${CMD} ${ADDR} --insecure
     sleep 2
 )
 done
